feat(v4): background workers and the health invariants
The verify throttle and the download slot manager, as tokio tasks. The slot
manager ranks candidates by the swarm's seeder count, which is why it had to
wait for the announcer: the engine knows connected peers, and a torrent nobody
is talking to yet reports none -- exactly the ones asking for a slot. Ranking
by that made the priority random.

The verify throttle ends once nothing is left to check. It is a boot-time job,
and a ticker running forever over a catalogue with nothing to verify is pure
cost at 244k torrents.

The health invariants are the other half. Each one is a statement that must
hold of a working engine, and each exists because something broke once and
nothing noticed:

- ghost catches an active torrent whose files are gone. Every received piece
  fails its hash and is re-requested forever, and thrown pieces are never
  counted, so it is invisible in total_download -- redl does not catch it. Only
  a stat does. This is the 80 GB one.
- starved catches a leecher whose swarm has seeds and who holds no peer: the
  shape of the left=0 announce bug, where the tracker withheld the peer list
  because we had called ourselves a seed.
- dual_seed catches the same hash seeded by both engines, which splits demand
  across two of our own peers for credit that counts the maximum, not the sum.

fake_seed takes its byte count from the caller rather than reading it back:
completion is derived from the piece picker, and walking every picker a second
time per scan is a quarter of a million pickers for a number already in hand.

+++ typhon-engine/src/hydra/announce/runner.rs
@@ -158,8 +158,8 @@ async fn announce_one(
                     cache.record(
                         &job.info_hash,
                         Entry {
-                            complete: resp.complete,
-                            incomplete: resp.incomplete,
+                            complete: resp.complete as i64,
+                            incomplete: resp.incomplete as i64,
                             tracker: tracker_url.clone(),
                             at: std::time::Instant::now(),
                             interval,
